AI/requirements.txt
torch
diffusers==0.18
transformers
scipy
ftfy
accelerate
tqdm
numpy
pyngrok
uvicorn
optimum-intel[openvino]
safetensors
invisible-watermark>=0.2.0